move App to it's own directory
