+++ src/app/AppRoutes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {BrowserRouter, Route, Switch} from 'react-router-dom';
-import { HomePage } from './home/HomePage';
-import {Recipes} from './recipes/pages/Recipes';
+import { HomePage } from '../home/HomePage';
+import {Recipes} from '../recipes/pages/Recipes';
 
 export const AppRoutes: React.FC = () => {
   return (
